architecture update
I'm still experimenting for chaining the layers, ignore the call function for now. But generally, I've found that using just dense for the actor and a combination of lstm with dense for critic to work the best. Also I don't see noticeable difference between embedding with one hot, so might be sticking with the simpler one hot for now.

also, I no longer use multiple learning rates